+++ tilegen/src/lib.rs
@@ -1,7 +1,12 @@
 mod regions;
 mod tiles;
 
-pub use regions::*;
+pub use regions::dir::*;
+pub use regions::line::*;
+pub use regions::point::*;
+pub use regions::rect::*;
+pub use regions::region::*;
+
 pub use tiles::context::TileContext;
 pub use tiles::grid::*;
 pub use tiles::replacement::*;
