+++ tilegen/src/regions/mod.rs
@@ -1,2 +1,5 @@
+pub mod dir;
+pub mod line;
 pub mod point;
+pub mod rect;
 pub mod region;
